Make several Team fields optional (#24)
Each of these fields was missing for me causing serde to panic
when listing pull requests on a specific private repo.
It appears that requested reviewers for a PR included a team that had a
parent team that did not have these fields in the API response.

Fwiw, I made no attempt to refer to Github API documentation,
  I merely ran a short snippet, saw serde panic on a serializing a missing
  field, made it optional, and repeated until my tool ran to completion.