allow fields in `PullRequest` and `Repository` to be empty
Fixes #125

As discussed in the PR, the `PullRequest` response in `list_runs`
does not match with the type that expects a lot of fields to not be
empty. This PR fixes by that wrapping those fields with `Option` and
thus making them optional.

As `PullRequest` also contains a field `repo` of type `Repository`
whose response also only includes only some of those fields (thus
causing error when serializing), this PR also fixes that by making
even those fields optional by wrapping them with `Option`.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>