Allow deserializing IDs as strings.
Fixes #123. It turns out that the notifications API returns
NotificationIDs in JSON as strings, not integers (e.g.: `id1: "1234"`).
See for example:
https://docs.github.com/en/rest/reference/activity#list-notifications-for-the-authenticated-user--code-samples

This PR adds an explicit Deserializer for each ID type that allows
either raw integers or string-encoded integers.

(We could have done this just for NotificationID, but it seemed easier
to just do for every ID type, and perhaps will help future-proof this
library.)