Add a function that exposes the installation token (#210)
You can use the installation token to access git repos using a regular git
client, so it's useful to expose this beyong the octocrab client once it
has been generated.

I've added a `installation_and_token` function which calls `installation` to
create an `Octocrab` instance for the installation and also forces the cache
of the installation token to be populated then returns both. This means that
the user has to get the token at the same time as they're getting the
installation client. But also that you can't try and get an installation
token from a client that isn't configured for authorising as an installation.