Make `encoding` optional in `Content`
This seems to never be returned by the API for entries when listing a
directory.

The GitHub API docs
https://docs.github.com/en/rest/repos/contents#get-repository-content
are a bit unclear on this, the response schema lists "encoding" as
required, but omit the type information. The examples are consistent
with it only being present if the path requested is a file.