gists+examples: Extend `GistsHandler` through `star(...)`, `unstar(...)`, `is_starred(...)` (#330)
* gists+examples: Extend `GistsHandler` by adding `star(...)`

The `star(...)` function takes a `gist_id` and 'stars' it. Has no
observable effect on the gist if it is already starred.

* gists+examples: Extend `GistsHandler` by adding `unstar(...)`

The `unstar(...)` function takes a `gist_id` and removes a 'star' from
it. Has no observable effect on the gist if it is not starred.

* examples: Rename star_a_gist -> star_unstar_a_gist

* gists: Add function to check if a gist is starred

Create `GistsHandler::is_starred`. This acceepts a `gist_id` and
returns if a gist is starred.

* examples: Extend the `star_unstar_a_gist` example

This fully demonstrates how to:

- Check if a gist is starred
- Star a gist
- Unstar a gist