examples/graphql_issues.rs: Use graphql_client to get stronger typing
This results in nicer code. The biggest downside is that we need the
GitHub GraphQL schema. We could download it in a build script, but it's
only about a megabyte, and changes rarely, so we can probably get away
with versioning a copy of it in the repo. It's nice to avoid build
scripts.