Fix personal token auth for pagination
Previously the AuthHeaderLayer added the auth header only when the
authority was empty (i.e., base URI for GitHub was supposed to be used).
However, the page URIs (e.g., next page) returned by GitHub API contains
the absolute URL including the hostname. Because of this, the following
code didn't work for resources that needed authentication (e.g., private
repos)

```
let page = /* get a page of something using octocrab */

// This fails with NOT_FOUND
octocrab.get_page(&page.next).await;
```

This fix adds the base URI to the AuthHeaderLayer so that it can check
whether a request is destined for GitHub (either the hostname is empty or
it is equal to the base URI).