error: don't include unconditional backtrace in Display impl
Users of this crate cannot currently get an error description without
the backtrace included. This may be undesirable for a number of
reasons, and is inconsistent with the examples in the `snafu`
crate (whose Display implementations generally do not include a
backtrace). Users who want a backtrace to be printed can still do so
for an octocrab Error `e` by writing:

    snafu::ErrorCompat::backtrace(&e)